chore: rename files and routes
