+++ packages/server/routes/api/moderator/middleware.ts
